ДЗ8. Добавлены тесты для стартера и основного приложения, application-test.properties.
