--- animal-creation-starter/src/main/java/ru/mts/hw_7/factory/ConcreteAnimalFactory.java
@@ -1,10 +1,0 @@
-package ru.mts.hw_7.factory;
-
-import ru.mts.hw_7.animal.Animal;
-
-import java.math.BigDecimal;
-import java.time.LocalDate;
-
-public interface ConcreteAnimalFactory {
-    Animal createAnimal(String name, String breed, BigDecimal cost, LocalDate birthDate);
-}
