+++ animal-creation-starter/src/main/java/ru/mts/hw_8/factory/AnimalTypes.java
@@ -1,0 +1,8 @@
+package ru.mts.hw_8.factory;
+
+public enum AnimalTypes {
+    WOLF,
+    SHARK,
+    RABBIT,
+    DEER
+}
