+++ animal-creation-starter/src/test/resources/application-test.properties
@@ -1,0 +1,4 @@
+animal.wolf.names=AntoshkaTest,OlejaTest
+animal.shark.names=DeniskaTest,VladnoTest
+animal.rabbit.names=EgorkaTest,NikitosTest
+animal.deer.names=SvenTest,OlafTest
