+++ build.gradle
@@ -17,6 +17,7 @@ dependencies {
     implementation 'javax.annotation:javax.annotation-api:1.3.2'
     implementation 'org.springframework.boot:spring-boot-starter'
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
+//    testImplementation 'junit:junit:4.12'
     implementation 'org.quartz-scheduler:quartz:2.3.2'
     compileOnly 'org.projectlombok:lombok:1.18.30'
     implementation project(':animal-creation-starter')
