+++ src/main/resources/application.properties
@@ -1,0 +1,4 @@
+animal.wolf.names=Antoshka,Oleja
+animal.shark.names=Deniska,Vladno
+animal.rabbit.names=Egorka,Nikitos
+animal.deer.names=Sven,Olaf
